a bunch of shit read description
so i streamed development of this cause i was bored
i added
2 players(most functionalities are present)
randomly spawning enemies that never touch eachother
you can shoot enemies
enemies have 3 health
enemies die and are removed upon death
bullets die and are removed upon death
memory is pretty decent at the moment, might need to revisit

Main note though, lot of this stuff is hardcoded. we need to revise it for when we want to properly do this, cause this is just a tech demo for review 2.